+++ changelog.md
@@ -6,6 +6,7 @@ Codion Change Log
 - UserPreferences.file() and delete() no longer accept empty string for filename.
 ### is.codion.swing.framework.ui
 - EntityTablePanelPreferences bug fixed, no longer applies empty preferences.
+- EntityApplicationPanel.Builder.frameSupplier() renamed frame().
 
 ## 0.18.41
 ## is.codion